refactor: Remove circular dependency in SharedUtils
