+++ src/aux-server/aux-web/aux-projector/interaction/DragOperation/BuilderFileDragOperation.ts
@@ -1,24 +1,11 @@
 import { Intersection, Vector3 } from 'three';
 import { Physics } from '../../../shared/scene/Physics';
-import {
-    File,
-    Workspace,
-    DEFAULT_WORKSPACE_SCALE,
-    fileRemoved,
-    fileUpdated,
-} from '@casual-simulation/aux-common/Files';
-import { keys } from 'lodash';
-import { gridPosToRealPos, Axial, posToKey } from '../../../shared/scene/hex';
+import { File } from '@casual-simulation/aux-common/Files';
 import {
     FileCalculationContext,
-    getContextMinimized,
-    getContextSize,
-    getBuilderContextGrid,
     isMinimized,
-} from '@casual-simulation/aux-common/Files/FileCalculations';
-import { ContextGroup3D } from '../../../shared/scene/ContextGroup3D';
+} from '@casual-simulation/aux-common';
 import { BuilderGroup3D } from '../../../shared/scene/BuilderGroup3D';
-import BuilderGameView from '../../BuilderGameView/BuilderGameView';
 import { BuilderInteractionManager } from '../BuilderInteractionManager';
 import { BaseBuilderFileDragOperation } from './BaseBuilderFileDragOperation';
 import { Simulation3D } from '../../../shared/scene/Simulation3D';
